organize mock logger and link mock logger to pip module
